move logic to setup/test-file-paths
